Modified

+++ src/Introductory Problems/Missing Number.cpp
@@ -3,9 +3,9 @@ using namespace std;
 
 int main()
 {
-	int n;
+	long long n;
 	cin >> n;
-	int sum = n*(n+1)/2;
+	long long sum = n*(n+1)/2;
 	for ( int i = 0 ; i < n - 1 ; i++ )
 	{
 		int a;
